requirements.txt
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<3.0.0
scikit-learn>=1.3.0,<2.0.0
matplotlib>=3.7.0,<4.0.0
seaborn>=0.12.0,<1.0.0
jupyter>=1.0.0
notebook>=7.0.0

# API Development
fastapi>=0.104.0,<1.0.0
uvicorn[standard]>=0.24.0,<1.0.0
pydantic>=2.0.0,<3.0.0

# Testing
pytest>=7.4.0,<8.0.0
pytest-cov>=4.1.0,<5.0.0
pytest-asyncio>=0.21.0,<1.0.0
httpx>=0.25.0,<1.0.0

# Code Quality
flake8>=6.1.0,<7.0.0
black>=23.11.0,<24.0.0
mypy>=1.7.0,<2.0.0

# CLI
click>=8.1.0,<9.0.0

# Model Management
joblib>=1.3.0,<2.0.0
mlflow>=2.8.0,<3.0.0

# Utilities
python-dotenv>=1.0.0,<2.0.0
pyyaml>=6.0.0,<7.0.0
